+++ examples/transform_translation.rs
@@ -1,7 +1,9 @@
-use bevy::prelude::*;
+use bevy::{color::palettes::css::*, prelude::*};
 use bevy_inspector_egui::{prelude::*, quick::ResourceInspectorPlugin};
 
 use bevy_tweening::{lens::*, *};
+
+mod utils;
 
 fn main() {
     App::default()
@@ -15,7 +17,7 @@ fn main() {
             ..default()
         }))
         .init_resource::<Options>()
-        .add_systems(Update, bevy::window::close_on_esc)
+        .add_systems(Update, utils::close_on_esc)
         .add_plugins(TweeningPlugin)
         .add_plugins(ResourceInspectorPlugin::<Options>::default())
         .add_systems(Startup, setup)
@@ -91,7 +93,7 @@ fn setup(mut commands: Commands) {
         commands.spawn((
             SpriteBundle {
                 sprite: Sprite {
-                    color: Color::RED,
+                    color: RED.into(),
                     custom_size: Some(Vec2::new(size, size)),
                     ..default()
                 },